+++ lib/sys/unix/sys/filesystem/constants.rb
@@ -25,6 +25,8 @@ module Sys
       MNT_DEFWRITE    = 0x02000000 # filesystem should defer writes
       MNT_MULTILABEL  = 0x04000000 # MAC support for individual labels
       MNT_NOATIME     = 0x10000000 # disable update of file access time
+      MNT_NOCLUSTERR  = 0x40000000 # disable cluster read
+      MNT_NOCLUSTERW  = 0x80000000 # disable cluster write
 
       MNT_VISFLAGMASK = (
         MNT_RDONLY | MNT_SYNCHRONOUS | MNT_NOEXEC |