+++ lib/sys/unix/sys/filesystem/structs.rb
@@ -6,6 +6,20 @@ require 'rbconfig'
 module Sys
   class Filesystem
     module Structs
+      # Used by DragonFlyBSD
+      class UUID < FFI::Struct
+        UUID_NODE_LEN = 6
+
+        layout(
+          :time_low, :uint32,
+          :time_mid, :uint16,
+          :time_hi_and_version, :uint16,
+          :clock_seq_hi_and_reserved, :uint8,
+          :clock_seq_low, :uint8,
+          :node, [:uint8, UUID_NODE_LEN]
+        )
+      end
+
       # The Statfs struct is a subclass of FFI::Struct that corresponds to a struct statfs.
       class Statfs < FFI::Struct
         # Private method that will determine the layout of the struct on Linux.
@@ -86,6 +100,31 @@ module Sys
                 :f_spare, [:ulong, 4]
               )
             end
+          when /dragonfly/i
+            layout(
+              :f_spare2, :long,
+              :f_bsize, :long,
+              :f_iosize, :long,
+              :f_blocks, :long,
+              :f_bfree, :long,
+              :f_bavail, :long,
+              :f_files, :long,
+              :f_ffree, :long,
+              :f_fsid, [:int32_t, 2],
+              :f_owner, :uid_t,
+              :f_type, :int,
+              :f_flags, :int,
+              :f_syncwrites, :long,
+              :f_asyncwrites, :long,
+              :f_fstypename, [:char, 16],
+              :f_mntonname, [:char, 80],
+              :f_syncreads, :long,
+              :f_asyncreads, :long,
+              :f_spares1, :short,
+              :f_mntfromname, [:char, 80],
+              :f_spares2, :short,
+              :f_spare, [:long,2]
+            )
           else
             layout(
               :f_bsize, :uint32,
@@ -148,22 +187,27 @@ module Sys
             :f_fsid, :ulong,
             :f_namemax, :ulong
           )
-        elsif RbConfig::CONFIG['host'] =~ /sunos|solaris/i
+        elsif RbConfig::CONFIG['host'] =~ /dragonfly/i
           layout(
             :f_bsize, :ulong,
             :f_frsize, :ulong,
-            :f_blocks, :uint64_t,
-            :f_bfree, :uint64_t,
-            :f_bavail, :uint64_t,
-            :f_files, :uint64_t,
-            :f_ffree, :uint64_t,
-            :f_favail, :uint64_t,
+            :f_blocks, :uint64,
+            :f_bfree, :uint64,
+            :f_bavail, :uint64,
+            :f_files, :uint64,
+            :f_ffree, :uint64,
+            :f_favail, :uint64,
             :f_fsid, :ulong,
-            :f_basetype, [:char, 16],
             :f_flag, :ulong,
             :f_namemax, :ulong,
-            :f_fstr, [:char, 32],
-            :f_filler, [:ulong, 16]
+            :f_owner, :uid_t,
+            :f_type, :uint,
+            :f_syncreads, :uint64,
+            :f_syncwrites, :uint64,
+            :f_asyncreads, :uint64,
+            :f_asyncwrites, :uint64,
+            :f_fsid_uuid, UUID,
+            :f_uid_uuid, UUID
           )
         elsif !linux64?
           layout(
@@ -197,17 +241,6 @@ module Sys
             :f_spare, [:int, 6]
           )
         end
-      end
-
-      # The Mnttab struct represents struct mnnttab from sys/mnttab.h on Solaris.
-      class Mnttab < FFI::Struct
-        layout(
-          :mnt_special, :string,
-          :mnt_mountp, :string,
-          :mnt_fstype, :string,
-          :mnt_mntopts, :string,
-          :mnt_time, :string
-        )
       end
 
       # The Mntent struct represents struct mntent from sys/mount.h on Unix.